job-tracker/requirements.txt
google-genai
requests
beautifulsoup4
python-dotenv
typer
rich
